dear loqor,
i've literally walked you through fixing it, as 3 out of 5 files you've committed are fixes that i've told you about in the gc.
the other 2 are fixes of your own rendering code that has nothing to do with me.
so technically
its your code that is bad
because you've been doing all that since the february
and im here like what
3rd day
not sincerely,
- theo